requirements.txt
pandas==2.1.3
numpy==1.26.0
pillow==9.5.0
scikit-learn==1.3.2
xgboost==2.0.2
flask==3.0.0
pillow==9.5.0